requirements.txt
google-genai==1.60.0
google-generativeai==0.8.6
requests==2.32.5
joblib==1.5.3
streamlit==1.53.1
markdown==3.10.1
yfinance>=0.2.0
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0